# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MC Roleplay Server API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members/{id}:
    get:
      operationId: getPublicProfile
      tags: [users]
      summary: Get public member profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUser"
  /members/{id}/followers:
    get:
      operationId: getPublicProfileFollowers
      tags: [users]
      summary: Get followers of a public member profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public member followers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicUser"
  /members/{id}/following:
    get:
      operationId: getPublicProfileFollowing
      tags: [users]
      summary: Get accounts followed by a public member profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public member following
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicUser"
...
  /follows:
    post:
      operationId: followUser
      tags: [follows]
      summary: Follow a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowInput"
      responses:
        "201":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
  /follows/{userId}:
    delete:
      operationId: unfollowUser
      tags: [follows]
      summary: Unfollow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unfollowed
...
  /admin/follows:
    post:
      operationId: adminCreateFollow
      tags: [follows]
      summary: Force a follow between two users (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFollowInput"
      responses:
        "201":
          description: Follow created
  /admin/bulk-followers:
    post:
      operationId: adminBulkCreateFollowers
      tags: [follows]
      summary: Generate N bot accounts and auto-follow a target user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBulkFollowInput"
      responses:
        "201":
          description: Bulk followers created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBulkFollowResult"
...
components:
  schemas:
...
    PublicUser:
      type: object
      required: [id, username, userTag, createdAt, isFollowing, followerCount, followingCount]
      properties:
        id:
          type: integer
        username:
          type: string
        userTag:
          type: string
          pattern: "^#[0-9]{3,}$"
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        youtubeLiveUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        isFollowing:
          type: boolean
        followerCount:
          type: integer
        followingCount:
          type: integer
    FollowInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
    AdminBulkFollowInput:
      type: object
      required: [targetUserId, count]
      properties:
        targetUserId:
          type: integer
        count:
          type: integer
          minimum: 1
          maximum: 10000
    AdminBulkFollowResult:
      type: object
      required: [botsCreated, followsCreated]
      properties:
        botsCreated:
          type: integer
        followsCreated:
          type: integer
...
    AdminFollowInput:
      type: object
      required: [followerId, followingId]
      properties:
        followerId:
          type: integer
        followingId:
          type: integer
    FollowResult:
      type: object
      required: [followerId, followingId]
      properties:
        followerId:
          type: integer
        followingId:
          type: integer
...