# lib/api-zod/src/generated/types/publicUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MC Roleplay Server API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicUser {
  id: number;
  username: string;
  /** @pattern ^#[0-9]{3,}$ */
  userTag: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  youtubeLiveUrl?: string | null;
  createdAt: string;
  isFollowing: boolean;
  followerCount: number;
  followingCount: number;
}